add snackbar

+++ src/app/notes/note/note.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Note } from '../Models/Notes';
 import { NotesService } from '../services/notes.service';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import {MatDialog} from '@angular/material/dialog';
+
 import * as moment from 'moment';
 moment.locale('es');
 @Component({
@@ -12,7 +15,7 @@ export class NoteComponent implements OnInit {
 
   @Input() note!: Note;
 
-  constructor(private notesService: NotesService) {
+  constructor(private notesService: NotesService, private _snackBar: MatSnackBar) {
   }
 
   ngOnInit(): void {
@@ -20,15 +23,31 @@ export class NoteComponent implements OnInit {
   }
 
   public onToggleArchiveNote(note: Note) {
-    note.archive ? this.notesService.desArchiveNote(note) : this.notesService.archiveNote(note);
+    if (note.archive) {
+      this.notesService.desArchiveNote(note);
+      this.openSnackBar('Nota Des-archivada !');
+    } else {
+      this.notesService.archiveNote(note);
+      this.openSnackBar('Nota Archivada !');
+    }
   }
 
   public onRemoveNote(note: Note) {
-    this.notesService.removeNote(note);
+    let confi = confirm('Deseas eliminar esta nota?')
+    if (confi) {
+      this.notesService.removeNote(note);
+      this.openSnackBar('Nota Eliminada !');
+    }
   }
 
   public onDateNote(date: Date) {
     return moment(date).calendar();
   }
 
+  private openSnackBar(message: string) {
+    this._snackBar.open(message, 'Cerrar');
+  }
+
 }
+
+
